+++ apps/forms/scaffolds/analytics/world/live-world-analytics.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import { fmtnum } from "@/scaffolds/analytics/stats";
-import { ResponseFeedProvider, useEditorState } from "@/scaffolds/editor";
+import { useEditorState } from "@/scaffolds/editor";
+import { ResponseFeedProvider } from "@/scaffolds/editor/feed";
 import { MapGL } from "@/theme/templates/formstart/default/mapgl";
 import React, { useEffect, useMemo, useState } from "react";
 import { MapProvider, useMap } from "react-map-gl";
